+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Removed `helpOption` from `ParsedCommand` type
 - Added standalone options
 - Added global options, completely redesigned help logic
+- Added `argv` and `initOptions` to the parsed object
 
 ## v2.1.0 2024-07-18